♻️ Refactor: ReadStatus 관련 로직 및 명칭을 ChannelMember로 전면 수정
- `ReadStatus` → `ChannelMember`로 클래스, 메서드 및 패키지 모듈 명칭 변경
- 관련 DTO, Controller, Service, Facade 등 전반적인 로직에 반영
- 불필요한 import 제거 및 패키지 구조 개선
