+++ src/main/java/com/sprint/mission/discodeit/dto/channelmember/request/ReadStatusCreateReq.java
@@ -1,0 +1,10 @@
+package com.sprint.mission.discodeit.dto.channelmember.request;
+
+import java.util.UUID;
+
+public record ReadStatusCreateReq(
+    UUID userId,
+    UUID channelId
+) {
+
+}
