--- src/main/java/com/sprint/mission/discodeit/dto/readstatus/request/ReadStatusCreateReq.java
@@ -1,8 +1,0 @@
-package com.sprint.mission.discodeit.dto.readstatus.request;
-
-import java.util.UUID;
-
-public record ReadStatusCreateReq(
-        UUID userId,
-        UUID channelId
-) {}
